requirements.txt
# requirements.txt
# Dependencias del proyecto IMDB Scraper

# Web scraping
beautifulsoup4==4.12.2
requests==2.31.0

# Scrapy para scraping avanzado
scrapy==2.11.1

# Async y red
aiohttp==3.9.3
httpx==0.27.0
tenacity==8.2.3  # Retry/backoff

# Base de datos
SQLAlchemy==2.0.41
asyncpg==0.29.0


# Utilidades
python-dotenv==1.0.1
colorlog==6.8.2

# Dashboards / export
Jinja2==3.1.3

# Tipado
typing-extensions==4.12.1
